demo new activation

+++ src/main/java/com/gomobile/integratedService/model/User.java
@@ -17,7 +17,7 @@ public class User {
     @Indexed(unique = true)
     private String email;
     // This need to be Double!
-    private Integer credit;
+    private double credit;
     private List<Order> orders;
 
     public User(String uid, String email, Integer credit) {
